IGNITE-20704 Add methods to fsync files and directries
