+++ modules/cli/src/main/java/org/apache/ignite/internal/cli/config/ini/IniConfigManager.java
@@ -46,8 +46,8 @@ import org.apache.ignite.internal.cli.config.exception.ConfigInitializationExcep
 import org.apache.ignite.internal.cli.config.exception.ProfileNotFoundException;
 import org.apache.ignite.internal.cli.core.exception.IgniteCliException;
 import org.apache.ignite.internal.cli.logger.CliLoggers;
-import org.apache.ignite.internal.cli.util.OperatingSystem;
 import org.apache.ignite.internal.logger.IgniteLogger;
+import org.apache.ignite.internal.util.OperatingSystem;
 
 /**
  * Implementation of {@link ConfigManager} based on {@link IniFile}.
