+++ modules/cli/src/testFixtures/java/org/apache/ignite/internal/cli/commands/cliconfig/TestConfigManagerHelper.java
@@ -32,7 +32,7 @@ import java.nio.file.Files;
 import java.nio.file.attribute.PosixFilePermission;
 import java.util.Set;
 import org.apache.ignite.internal.cli.config.ConfigManager;
-import org.apache.ignite.internal.cli.util.OperatingSystem;
+import org.apache.ignite.internal.util.OperatingSystem;
 
 /**
  * Test factory for {@link ConfigManager}.
