+++ modules/core/src/main/java/org/apache/ignite/internal/util/OperatingSystem.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.util;
+package org.apache.ignite.internal.util;
 
 import java.util.Locale;
